v4.0

+++ springboot/ps5GameTradingSystem/src/main/java/com/zxy/ps5gametradingsystem/entity/Recycle.java
@@ -10,7 +10,7 @@ public class Recycle {
     private  String id;
     private  String userId;
     private  String gameName;
-    private  Integer valuation;
+    private  Double valuation;
     private  String recyclingStatus;
     private  String version;
     private  String language;
